Controllers added

+++ Logistics/Logistics.Core/Entities/WeightRange.cs
@@ -11,6 +11,6 @@ namespace Logistics.Core.Entities
         public Guid Id { get; set; }
         public float MinimalWeight { get; set; }
         public float MaximalWeight { get; set; }
-        public float PriceCoefficient { get; set; }
+        public double PriceCoefficient { get; set; }
     }
 }
